training/requirements.txt
gymnasium
stable-baselines3
pymavlink
pyyaml
tensorboard
